Add base83 encoder impl
Based on the Go and C reference implementations. Structure
based on the zig standard library's base64 encoder. This
implementation is closer to the C one since it does not validate
the length given to the encoder, it will just use the length
given. The Go implementation checks the length and returns an
error if it finds it's invalid.

In the future may want to consider changing the function signatures
to use usize instead of i64. The use of i64 is mostly to match
what Go/C have, but that doesn't seem necessary since the algorithm
is lossy anyways with the wrapping on overflow.