Rewrite test cases
In Go it's common to have a slice of test cases, but in
zig there's no way to name tests inside a test dynamically
or statically through a name field. The output on a failing
test in these cases is difficult to know which test failed
so put each test on its own so that the stack trace will
point to the exact case that failed.